# `<regex>`: Speed up searches for regexes that start with assertions