--- a/stl/inc/regex
+++ b/stl/inc/regex
@@ -1741,7 +1741,7 @@ public:
         return true;
     }
 
-    _BidIt _Skip(_BidIt, _BidIt, _Node_base* = nullptr);
+    _BidIt _Skip(_BidIt, _BidIt, _Node_base* = nullptr, unsigned int _Recursion_depth = 0U);
 
 private:
     _Tgt_state_t<_It> _Tgt_state;
@@ -4051,10 +4051,12 @@ bool _Matcher2<_BidIt, _Elem, _RxTraits, _It, _Alloc>::_Match_pat(_Node_base* _N
 }
 
 template <class _BidIt, class _Elem, class _RxTraits, class _It, class _Alloc>
-_BidIt _Matcher2<_BidIt, _Elem, _RxTraits, _It, _Alloc>::_Skip(_BidIt _First_arg, _BidIt _Last, _Node_base* _Node_arg) {
+_BidIt _Matcher2<_BidIt, _Elem, _RxTraits, _It, _Alloc>::_Skip(
+    _BidIt _First_arg, _BidIt _Last, _Node_base* _Node_arg, unsigned int _Recursion_depth) {
     // skip until possible match
     // assumes --_First_arg is valid
-    _Node_base* _Nx = _Node_arg ? _Node_arg : _Rep;
+    constexpr unsigned int _Max_recursion_depth = 50U;
+    _Node_base* _Nx                             = _Node_arg ? _Node_arg : _Rep;
 
     while (_First_arg != _Last && _Nx) { // check current node
         switch (_Nx->_Kind) { // handle current node's type
@@ -4145,17 +4147,54 @@ _BidIt _Matcher2<_BidIt, _Elem, _RxTraits, _It, _Alloc>::_Skip(_BidIt _First_arg
                 break;
             }
 
+        case _N_assert:
+            {
+                if (_Recursion_depth >= _Max_recursion_depth) {
+                    return _First_arg;
+                }
+
+                _Node_assert* _Node = static_cast<_Node_assert*>(_Nx);
+                _First_arg          = _Skip(_First_arg, _Last, _Node->_Child);
+                _BidIt _Next;
+                for (;;) {
+                    _Next = _Skip(_First_arg, _Last, _Node->_Next, _Recursion_depth + 1U);
+                    if (_Next == _First_arg) {
+                        return _First_arg;
+                    }
+
+                    _First_arg = _Skip(_Next, _Last, _Node->_Child, _Recursion_depth + 1U);
+                    if (_Next == _First_arg) {
+                        return _First_arg;
+                    }
+                }
+            }
+
+        case _N_neg_assert:
+            // we skip the negated assertion body and continue examining the rest of the regex
+            break;
+
+        case _N_wbound:
+            {
+                bool _Negated   = (_Nx->_Flags & _Fl_negate) != 0;
+                bool _Prev_word = _STD _Is_word(*_STD _Prev_iter(_First_arg));
+                for (; _First_arg != _Last; ++_First_arg) {
+                    bool _Next_word = _STD _Is_word(*_First_arg);
+                    if (_Negated == (_Next_word == _Prev_word)) {
+                        break;
+                    }
+                    _Prev_word = _Next_word;
+                }
+                return _First_arg;
+            }
+
         case _N_begin:
+        case _N_endif:
             break;
 
         case _N_end:
         case _N_none:
-        case _N_wbound:
         case _N_dot:
-        case _N_assert:
-        case _N_neg_assert:
         case _N_back:
-        case _N_endif:
         case _N_end_rep:
         default:
             return _First_arg;